fix(astronaut): remove currentspace from astronaut
Put the currentSpace subscriber in the affected code - one of the
next tasks is to remove this entirely and do it using observables.
However, to keep things working, we're doing it in stages.

In particular, I want to get rid of astronaut.getCurrentSpace() asap
so that no one else starts using it for new code! We don't want to
promote bad practices...

Also, moving this 'bad' code in to the services, makes reasoning
about them easier.